Improved the stereo audio system, but more importantly removed some redundant attributes to the Abstract class:
I literally didn't use Name ever and it was a mandatory attribute, so I got rid of it. It would have been pretty useful if I was making an editor interface like Unity or Godot, but tbh if you're doing everything through code you'll have a variable name for things. And you can just use a tag if you really need a string with a name for something

I also got rid of the Script attribute because the user writes their own process subroutine for each abstract, so there's no point having it around.

I also got rid of parent and children from abstract initialisation, but they're still there and can be used